+++ src/functions/security.py
@@ -22,7 +22,7 @@ def checkPerm(app, roles, perms):
         perms = [perms]
     for role in roles:
         for perm in perms:
-            if perm in app.config.__dict__["perms"].__dict__.keys() and role in app.config.__dict__["perms"].__dict__[perm]:
+            if perm in app.config_dict["perms"].keys() and role in app.config_dict["perms"][perm]:
                 return True
     return False
 
@@ -269,7 +269,7 @@ async def auth(authorization, request, allow_application_token = False, check_me
             ok = False
             for role in roles:
                 for perm in required_permission:
-                    if perm in app.config.__dict__["perms"].__dict__.keys() and role in app.config.__dict__["perms"].__dict__[perm] or role in app.config.__dict__["perms"].__dict__["administrator"]:
+                    if perm in app.config_dict["perms"].keys() and role in app.config_dict["perms"][perm] or role in app.config_dict["perms"]["administrator"]:
                         ok = True
 
             if not ok:
@@ -410,7 +410,7 @@ async def auth(authorization, request, allow_application_token = False, check_me
 
             for role in roles:
                 for perm in required_permission:
-                    if perm in app.config.__dict__["perms"].__dict__.keys() and role in app.config.__dict__["perms"].__dict__[perm] or role in app.config.__dict__["perms"].__dict__["administrator"]:
+                    if perm in app.config_dict["perms"].keys() and role in app.config_dict["perms"][perm] or role in app.config_dict["perms"]["administrator"]:
                         ok = True
 
             if not ok: